messanger-47: Добавление ack на клиенте и сервере + организация файлов и папок внутри бизнес-ядра на сервере и клиенте
