+++ app/src/core/socket/Socket.ts
@@ -4,9 +4,9 @@ import { io } from "socket.io-client";
 import SocketController from "@core/socket/SocketController";
 import Logger from "@service/Logger";
 import i18next from "@service/i18n";
-import { RECONECTION_ATTEMPTS, RECONNECTION_DELAY, API_URL } from "@utils/constants";
+import { SOCKET_RECONECTION_ATTEMPTS, SOCKET_RECONNECTION_DELAY, API_URL, SOCKET_ACK_TIMEOUT } from "@utils/constants";
 import { AppDispatch } from "@custom-types/redux.types";
-import { SocketType } from "@custom-types/socket.types";
+import { ClientToServerEvents, SocketType } from "@custom-types/socket.types";
 import { MainClientEvents, SocketEvents } from "@custom-types/events";
 import { IUser } from "@custom-types/models.types";
 
@@ -35,8 +35,8 @@ export default class Socket extends EventEmitter {
         this._socket = io(API_URL, { 
             transports: ["websocket"],                  // Виды транспортов (идут один за другим по приоритету, данный массив на сервере должен совпадать)
             autoConnect: true,                          // Автоматически подключаться при создании экземпляра клиента (без вызова connect())
-            reconnectionAttempts: RECONECTION_ATTEMPTS, // Количество попытокпереподключения перед тем, как закрыть соединение
-            reconnectionDelay: RECONNECTION_DELAY,      // Задержка между попытками переподключения
+            reconnectionAttempts: SOCKET_RECONECTION_ATTEMPTS, // Количество попыток переподключения перед тем, как закрыть соединение
+            reconnectionDelay: SOCKET_RECONNECTION_DELAY,      // Задержка между попытками переподключения
             forceNew: false,                            // Не создает новое соединение, если оно уже существует
             upgrade: true,                              // Позволяет улучшать соединение с polling до websocket (в нашем случае улучшается с первого http-запроса (handshake) до websocket)
             closeOnBeforeunload: true,                  // Позволяет сокет соединению автоматически закрываться при событии beforeunload (закрытие вкладки/браузера/обновление страницы)
@@ -47,6 +47,23 @@ export default class Socket extends EventEmitter {
         this._socketController = new SocketController(this._socket, this._dispatch, this._me);
 
         this._bindSocketControllerListeners();
+    }
+
+    // Основной метод отправки события с клиента на сервер с добавлением ack (подтверждение обработки сервером данного события)
+    // Необходимо корректно указать тип аргументов => [infer _] нам необходим, но на него ругается линтер
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    async send<T extends keyof ClientToServerEvents>(type: T, ...args: Parameters<ClientToServerEvents[T]> extends [...infer R, infer _] ? R : any[]) {
+        try {
+            const { success, message, timestamp } = await this._socket.timeout(SOCKET_ACK_TIMEOUT).emitWithAck(type, ...args);
+
+            if (!success) {
+                throw i18next.t("core.socket.error.emit_event_on_the_server", { message, timestamp });
+            }
+
+            logger.debug(i18next.t("core.socket.emit_handled", { type, timestamp }));
+        } catch (error) {
+            this.emit(MainClientEvents.ERROR, i18next.t("core.socket.error.emit", { type, message: error }));
+        }
     }
 
     disconnect() {
