+++ app/src/locales/ru.json
@@ -68,8 +68,11 @@
 	"sign-up-module.repeat_password": "Повторите пароль",
 
 	"core.profiles-controller.error.profile_not_exists": "Профиля с id: {{id}} не существует.",
+
 	"core.socket.error.user_not_exists": "Объект пользователя не существует.",
 	"core.socket.error.connection": "Ошибка сокет-соединения [isSocketActive: {{isSocketActive}}]: {{message}}",
+	"core.socket.error.emit": "Ошибка при обработке события на сервере [type: {{type}}]: {{message}}",
+	"core.socket.error.emit_event_on_the_server": "[{{timestamp}}] | При обработке данного события на сервере возникла ошибка: {{message}}",
 
 	"core.catch-errors.error": "Ошибка",
 	"core.catch-errors.error.network": "Ошибка при получении ответа от сервера. Скорее всего сервер недоступен в данный момент. Проверьте интернет соединение.",
@@ -82,6 +85,7 @@
 	"core.socket.online_users": "Пользователи онлайн: ",
 	"core.socket.new_user_connected": "Новый пользователь подключился: ",
 	"core.socket.user_disconnected": "Пользователь отключился: ",
+	"core.socket.emit_handled": "Событие обработано на сервере [type={{type}}, timestamp={{timestamp}}].",
 
 	"header.settings": "Настройки",
 	"header.help": "Помощь",
