+++ server/package.json
@@ -31,7 +31,7 @@
     "redis": "^4.7.0",
     "sequelize": "^6.37.1",
     "sharp": "^0.33.4",
-    "socket.io": "^4.7.5",
+    "socket.io": "^4.8.1",
     "tedious": "^18.1.0",
     "uuid": "^9.0.1",
     "winston": "^3.17.0"
