+++ server/types/events.ts
@@ -1,4 +1,6 @@
 export enum SocketEvents {
+    SEND = "send",
+    SEND_BROADCAST = "sendBroadcast",
     NOTIFY_ANOTHER_USER = "notifyAnotherUser",
     NOTIFY_ALL_ANOTHER_USERS = "notifyAllAnotherUsers",
     HANDLE_ERROR = "handleError"
